+++ hefquin-access/src/main/java/se/liu/ida/hefquin/federation/access/impl/reqproc/SPARQLRequestProcessorImpl.java
@@ -144,40 +144,4 @@ public class SPARQLRequestProcessorImpl implements SPARQLRequestProcessor
 
 	} // end of MySolutionConsumer
 
-	/**
-	 * Determines whether it is semantically safe to override the projection
-	 * of the given query.
-	 *
-	 * <p>Overriding the projection means replacing the SELECT clause with a
-	 * new set of variables (e.g., for projection pushdown). This is only safe
-	 * for "simple" SELECT queries without features that depend on the original
-	 * projection.</p>
-	 *
-	 * <p>In particular, projection must not be overridden if the query contains:
-	 * <ul>
-	 *   <li>GROUP BY (projection affects grouping semantics)</li>
-	 *   <li>HAVING clauses (depends on grouped results)</li>
-	 *   <li>Aggregators (e.g., COUNT, SUM), which rely on specific projection expressions</li>
-	 * </ul>
-	 * </p>
-	 *
-	 * @param q the query to inspect
-	 * @return {@code true} if projection can be safely overridden; {@code false} otherwise
-	 */
-	protected static boolean isSafeToOverrideProjection(final Query q)
-	{
-		// Query must be SELECT type
-		if ( ! q.isSelectType() ) return false;
-
-		// Must not contain GROUP BY
-		if ( q.hasGroupBy() ) return false;
-
-		// Must not contain HAVING (optional but safe)
-		if ( q.hasHaving() ) return false;
-
-		// Must not have expressions in SELECT
-		if ( q.hasAggregators() ) return false;
-
-		return true;
-	}
 }
